+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# Change Log
+# Angular Follow Selector changelog
 
-## 1.0.0
-- Initial release of Angular Follow Selector extension
+## 1.1.0 (2017-06-09)
+- Added a new feature that lets you open the template and styling files directly from the component.
+
+## 1.0.0 (2017-05-21)
+- Initial release of Angular Follow Selector extension.